refactor(otlp-exporter-base): remove exports that were meant to be internal
