+++ experimental/packages/otlp-exporter-base/test/common/bounded-queue-export-promise-handler.test.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { createBoundedQueueExportPromiseHandler } from '../../src';
+import { createBoundedQueueExportPromiseHandler } from '../../src/bounded-queue-export-promise-handler';
 import * as assert from 'assert';
 
 describe('BoundedQueueExportPromiseHandler', function () {
